Event Ids are now long instead of int. Updated dependencies. Updated documents as well. Created a new consolidated migration.

+++ TicketManager/TicketManager.DataAccess.Documents/DataModel/ChangeTrackedObjectBase.cs
@@ -8,6 +8,6 @@ namespace TicketManager.DataAccess.Documents.DataModel
 
         public string LastChangedBy { get; set; }
 
-        public int LastKnownChangeId { get; set; }
+        public long LastKnownChangeId { get; set; }
     }
 }
