+++ TicketManager/TicketManager.DataAccess.Events/DataModel/TicketCommentPostedEvent.cs
@@ -4,7 +4,7 @@ namespace TicketManager.DataAccess.Events.DataModel
 {
     public class TicketCommentPostedEvent : EventBase, ITicketEvent
     {
-        public int TicketCreatedEventId { get; set; }
+        public long TicketCreatedEventId { get; set; }
 
         [ForeignKey(nameof(TicketCreatedEventId))]
         public virtual TicketCreatedEvent TicketCreatedEvent { get; set; }
