+++ TicketManager/TicketManager.WebAPI/DTOs/Commands/AddTicketLinksCommand.cs
@@ -7,7 +7,7 @@ namespace TicketManager.WebAPI.DTOs.Commands
     public class AddTicketLinksCommand : TicketCommandBase
     {
         [JsonConstructor]
-        public AddTicketLinksCommand(int ticketId, string raisedByUser, TicketLinkDTO[] links)
+        public AddTicketLinksCommand(long ticketId, string raisedByUser, TicketLinkDTO[] links)
             : base(ticketId, raisedByUser)
         {
             Links = links ?? Array.Empty<TicketLinkDTO>();
