+++ TicketManager/TicketManager.WebAPI/DTOs/Commands/PostCommentToTicketCommand.cs
@@ -3,10 +3,10 @@ using TicketManager.WebAPI.DTOs.Commands.Abstractions;
 
 namespace TicketManager.WebAPI.DTOs.Commands
 {
-    public class PostCommentToTicketCommand : TicketCommandBase<int>, ITicketCommand
+    public class PostCommentToTicketCommand : TicketCommandBase<long>, ITicketCommand
     {
         [JsonConstructor]
-        public PostCommentToTicketCommand(int ticketId, string raisedByUser, string commentText)
+        public PostCommentToTicketCommand(long ticketId, string raisedByUser, string commentText)
             : base(ticketId, raisedByUser)
         {
             CommentText = commentText;
